+++ .gitignore
@@ -60,9 +60,12 @@ venv/
 # [sass] Files generated by Sass
 *.css
 
-node_modules/
 bower_components/
 *.log
 
 build/
 dist/
+
+# Project-specific ignores
+.dotrun.json
+.venv/
