+++ package.json
@@ -3,21 +3,48 @@
   "license": "LGPL v3",
   "scripts": {
     "clean": "rm -rf node_modules yarn-error.log css static/css *.log *.sqlite _site/ build/ .jekyll-metadata .bundle",
-    "watch": "watch -p 'static/sass/**/*.scss' -c 'yarn run build'",
-    "build": "node-sass --include-path node_modules static/sass --source-map true --output-style compressed --output static/css && postcss --map false --use autoprefixer --replace 'static/css/**/*.css'",
+    "watch": "watch -p 'static/sass/**/*' -p 'static/js/**/*.js' -c 'yarn run build'",
+    "watch-js": "webpack --watch",
+    "watch-scss": "watch -p 'static/sass/**/*.scss' -c 'yarn run build-scss'",
+    "build": "yarn run build-css && yarn run build-js",
+    "build-css": "node-sass --include-path node_modules static/sass --source-map true --output-style compressed --output static/css && postcss --map false --use autoprefixer --replace 'static/css/**/*.css'",
+    "build-js": "yarn run build-js-bundle",
+    "build-js-bundle": "webpack",
     "format-python": "black --line-length 79 webapp",
     "lint-python": "flake8 webapp tests && black --check --line-length 79 webapp tests",
-    "lint-scss": "sass-lint static/**/*.scss --verbose --no-exit",
+    "lint-js": "eslint static/js",
+    "lint-scss": "stylelint static/**/*.scss",
     "serve": "./entrypoint 0.0.0.0:${PORT}",
+    "start": "yarn run build && concurrently --raw 'yarn run watch' 'yarn run serve'",
     "test": "yarn run lint-scss && yarn run lint-python && yarn run test-python",
     "test-python": "python3 -m unittest discover tests"
   },
   "dependencies": {
-    "autoprefixer": "9.7.6",
+    "@canonical/global-nav": "^2.4.1",
+    "autoprefixer": "9.7.5",
+    "babel-core": "^6.26.3",
+    "babel-eslint": "10.1.0",
+    "babel-loader": "^8.1.0",
+    "babel-preset-env": "^1.7.0",
+    "eslint": "6.8.0",
+    "eslint-config-prettier": "6.10.1",
+    "eslint-plugin-prettier": "3.1.2",
+    "expose-loader": "^0.7.5",
+    "husky": "4.2.3",
+    "lint-staged": "10.1.1",
     "node-sass": "4.13.1",
     "postcss-cli": "7.1.0",
-    "sass-lint": "1.13.1",
-    "vanilla-framework": "2.9.0",
-    "watch-cli": "0.2.3"
+    "prettier": "2.0.2",
+    "terser-webpack-plugin": "^2.3.5",
+    "vanilla-framework": "2.10.0",
+    "watch-cli": "0.2.3",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11"
+  },
+  "devDependencies": {
+    "concurrently": "5.1.0",
+    "stylelint": "13.2.1",
+    "stylelint-config-standard": "20.0.0",
+    "stylelint-order": "4.0.0"
   }
 }
