--- static/js/README.md
@@ -1,5 +1,0 @@
-# JS
-
-**This directory is not required, you can delete it if you don't want to use it.**
-
-This directory contains your JavaScript files.
