+++ static/js/global-nav.js
@@ -1,0 +1,3 @@
+import { createNav } from "@canonical/global-nav";
+
+createNav({ maxWidth: "72rem" });
