+++ templates/404.html
@@ -1,1 +1,20 @@
-404
+{% extends '_layout.html' %}
+
+{% block title %}404: Page not found{% endblock %}
+
+{% block meta_description %}MicroK8s | Sorry, we can't find the page you're looking for.{% endblock %}
+
+{% block content %}
+<div class="p-strip is-deep u-extra-space">
+  <div class="row">
+    <div class="col-6">
+      <img src="https://assets.ubuntu.com/v1/4de1c56c-404_v1.svg" alt="Error owl" width="400" height="400" />
+    </div>
+    <div class="col-6">
+      <h1>404: Page not found</h1>
+      <p>Sorry, we can't find the page you're looking for.</p>
+      <p>If you think this is an error on our site, please <a href="https://github.com/canonical-web-and-design/https://microk8s.io//issues/new" class="p-link--external">file a bug</a>.</p>
+    </div>
+  </div>
+</div>
+{% endblock content %}
