+++ templates/500.html
@@ -1,1 +1,25 @@
-500
+{% extends 'base_layout.html' %}
+
+{% block title %}500: Server error{% endblock %}
+
+{% block meta_description %}MicroK8s | Something&rsquo;s gone wrong.{% endblock %}
+
+{% block content %}
+<div class="p-strip is-deep u-extra-space">
+  <div class="row">
+    <div class="col-6">
+      <img src="https://assets.ubuntu.com/v1/4de1c56c-404_v1.svg" alt="Error owl" width="400" height="400" />
+    </div>
+    <div class="col-6">
+      <h1>500: Server error</h1>
+      <p class="p-heading--four">Something&rsquo;s gone wrong.</p>
+      {% if message %}<blockquote><code>{{ message }}</code></blockquote>{% endif %}
+      <p>
+        Try reloading the page.
+        If the error persists, please note that it may be a <a class="p-link--external" href="https://github.com/canonical-web-and-design/https://microk8s.io/issues">known issue</a>.
+        If not, please <a class="p-link--external" href="https://github.com/canonical-web-and-design/https://microk8s.io/issues/new">file a new issue</a>.
+      </p>
+    </div>
+  </div>
+</div>
+{% endblock content %}
