+++ templates/partial/_footer.html
@@ -1,0 +1,25 @@
+<footer>
+  <div class="p-strip">
+    <div class="u-fixed-width">
+      <p>
+        MicroK8s was built by the dedicated <a class="p-link--external" href="https://www.ubuntu.com/kubernetes">Kubernetes team at Canonical</a> for the developer community. We also make <a class="p-link--external" href="https://ubuntu.com/kubernetes/install#cluster">Charmed Kubernetes</a> for multi-cloud scalable clusters.
+      </p>
+      <p class="u-no-max-width">
+        © {{ now("%Y") }} Canonical Ltd. Ubuntu and Canonical are registered trademarks of Canonical&nbsp;Ltd.
+      </p>
+      <nav>
+        <ul class="p-inline-list--middot">
+          <li class="p-inline-list__item">
+            <a class="p-link--external" href="https://www.ubuntu.com/legal">Legal information</a>
+          </li>
+          <li class="p-inline-list__item">
+            <a class="p-footer__link p-link--external" href="https://github.com/canonical-websites/microk8s.io/issues/new">Report a bug on this site</a>
+          </li>
+        </ul>
+        <span class="u-off-screen">
+          <a href="#">Go to the top of the page</a>
+        </span>
+      </nav>
+    </div>
+  </div>
+</footer>
