+++ templates/partial/_navigation.html
@@ -1,0 +1,32 @@
+<header id="navigation" class="p-navigation">
+  <div class="p-navigation__row">
+    <div class="p-navigation__banner">
+      <div class="p-navigation__logo u-vertically-center">
+        <a class="p-navigation__link" href="/" style="color:white">
+          MicroK8s
+        </a>
+      </div>
+      <a href="#navigation" class="p-navigation__toggle--open" title="menu">Menu</a>
+      <a href="#navigation-closed" class="p-navigation__toggle--close" title="close menu">Close menu</a>
+    </div>
+    <nav class="p-navigation__nav">
+      <span class="u-off-screen">
+        <a href="#main-content">Jump to main content</a>
+      </span>
+      <ul class="p-navigation__links" role="menu">
+        <li class="p-navigation__link " role="menuitem">
+          <a href="/docs/">Docs</a>
+        </li>
+        <li class="p-navigation__link" role="menuitem">
+          <a class="p-link--external" href="https://discuss.kubernetes.io/tags/microk8s">Discourse</a>
+        </li>
+        <li class="p-navigation__link" role="menuitem">
+          <a class="p-link--external" href="https://github.com/ubuntu/microk8s">Code</a>
+        </li>
+        <li class="p-navigation__link" role="menuitem">
+          <a class="p-link--external" href="https://github.com/ubuntu/microk8s/issues">Bugs</a>
+        </li>
+      </ul>
+    </nav>
+  </div>
+</header>
